Conflicts resolved

+++ Suffix_tree.c
@@ -1,12 +1,12 @@
 #include<stdio.h>
 #include"Suffix_tree.h"
-long int edgeLength(struct SuffixTreeNode*a)
+int edgeLength(struct SuffixTreeNode*a)
 {
   long int e=*(a->end);
   long int s= a->start ; 
   return (e-s+1);
 }
-struct SuffixTreeNode *newNode(long int start, int *end)
+struct SuffixTreeNode *newNode(int start, int *end)
 {
    count = count + 1;
    struct SuffixTreeNode *STnode = (struct SuffixTreeNode *)malloc(sizeof(struct SuffixTreeNode));
